Use Vec<> instead of Option<> for loop state management
Previously we only stored the currently active loop state
of the NestedState. The old state was stored in a local
variable of the currently inner most loop.
This was sufficient because continue and break could
only affect the innermost loop.

For upcoming named loops with labeled continue and break
statements we need to be able to access all loop states
from the current state.