Add rational type with 'a:b' syntax.
In order to make that work without totally breaking
dicts/sets, ternaries, switch we also introduced
the ability to ignore specific tokens in precedence/expression parsing.

So for example these three constructs stop all expression parsing
at the first ':' token. These excluded tokens need to be propagated
to all other expressions that are parsed that are not "closed".

Otherwise "1 ? 2 or 3 : 4"  would parse this as '1 ? 2 or (3:4)'
which then fails.

"Closed" forms like lists, sets, dicts, grouping or rules that do not
parse arbitrary expressions do not need this.

For example '1 ? [2, 2:3] : 4' never leaks the ':' from the list back
into the ternary.

This same infrastructure will also have to be used when implementing
tuple via "," separation in order to not mess with lists, dicts, sets or
argument lists.