Add proper Result return types to allow shortcircuiting on errors
This is required because many dunder methods make use of the
`run_function` machinery which runs bytecode eagerly to use it
in a native function or for control flow (current uses).

However callers need to be able to short circuit all follow up
actions after such an error occured in order to not corrup the stack.