Add state to keep track of whether an exception is currently in flight
Specifically there is now a boolean that is true in the time
between "unwind" being called and the start of execution for the
next bytecode instruction, which would be the first exception
handler that is checked for a match.

This is needed because some instruction (is_falsey) or native functions
might manipulate the stack in the meantime, which could cause
an inconsistent state.